backend/requirements.txt
langchain==0.1.16
langchain-community==0.0.36
langchain-openai==0.1.3
openai==1.14.3

fastapi
uvicorn

python-dotenv
requests
